- add 3 power box in HUD

+++ Mario-game/PlayScene.cpp
@@ -96,7 +96,7 @@ void CPlayScene::_ParseSection_ANIMATIONS(string line)
 	LPANIMATION ani = new CAnimation();
 
 	int ani_id = atoi(tokens[0].c_str());
-	for (int i = 1; i < tokens.size(); i += 2)	// why i+=2 ?  sprite_id | frame_time  
+	for (int i = 1; (float)i < tokens.size(); i += 2)	// why i+=2 ?  sprite_id | frame_time  
 	{
 		int sprite_id = atoi(tokens[i].c_str());
 		int frame_time = atoi(tokens[i+1].c_str());
@@ -344,7 +344,7 @@ void CPlayScene::Render()
 	CMaps::GetInstance()->RenderBackground();
 	CMaps::GetInstance()->RenderShading();
 	CMaps::GetInstance()->RenderGraphic();
-	for (int i = 0; i < objects.size(); i++)
+	for (int i = 0; (float)i < objects.size(); i++)
 		objects[i]->Render();
 
 }
@@ -370,7 +370,7 @@ void CPlayScene::Clear()
 */
 void CPlayScene::Unload()
 {
-	for (int i = 0; i < objects.size(); i++)
+	for (int i = 0; (float)i < objects.size(); i++)
 		delete objects[i];
 
 	objects.clear();
